Fix regression in gameboy
My earlier change to ChangeIfTypeIsEquivalent introduced regression in gameboy benchmark. Earlier if equivalent type cache was full, we evicted existing type from cache to store new type we saw. In ... I changed it to say "if i see non-shared type and cache is full, then treat the type as non-equivalent and perform helper path to Ld/St operation". This leds to lot of repeating non-shared types to perform operation using helper path or we bailout. Changed the heuristics to evict non-shared type if we see non-shared type. This fix the regression.