Correct a flag added recently, fixing formatting of exceptions thrown in catch blocks.
I recently added a parameter to AutoCatchHandlers to help detection
of whether or not an exception can be caught in user code, to match
the -nonative output better. Unfortunately, I set it incorrectly in
one place; namely, the case where we're in a "catch" statement made
by the user. Flipping the flag the other way fixes the issue.