change page allocator cleanup heuristic
changing from idle for 10 seconds to 500 ms to trigger the cleanup
changing from waitable timer to threadpool timer. Because waitable time using scheduling thread's APC queue to trigger the callback, and looks RPC server is using thread pool, the waitable timer scheduling thread can be gone from the thread pool so the callback won't happen, if that's the case we'll miss a scheduled cleanup. Our cleaning up procedure is not thread affinity so it should be OK to use the default call back environment.