Module hangs if it imports an errored module
ResolveExternalModuleDependencies has no problem today resolving against a
module that has already been determined to be in an error state.  In this
case, it has already notified its dependents at the time of its failure,
and no further notification will occur.  As a result, the module hangs if
we do not handle this scenario.

This fix addresses the issue by not proceeding if the module being
imported is already in an error state.  The error is passed upward.