Retain slot capacity while type sharing
If we are sharing types, we should make sure the old type and new type has same slotCapacity. Added additional assert.

Also added a `DBG` only flag on `DynamicType` that will indicate if the type was cached for ChangePrototype feature. This will make debugging easy especially when one of the cached types are involved in failure in OOP codegen.