Fix JsParse/JsRun for JavascriptString argument
JavascriptString is internally wide, so pass correct length and flag in
CompileRun.

GetSz for JavascriptString can allocate memory and need to handle OOM so
need API wrapper.

We are using the wrapper twice but refactoring is not simple as
RunScriptCore is called at multiple, for now lets keep two wrappers