Fix crash due to label breaking precise captures
Precise closure capture breaks when a label exists with the same name as
a variable or function.

Label parsing used to rely on ParseExpr(), checking after the fact to
see if the expression was a single identifier and if it is followed by a
tkColon. Since it parsed, however, a pidref was pushed for the
identifier setting up a new reference to that identifier which could
then trick the front-end into thinking there was a closure capture where
there actually wasn't.  This leads to invalid StSlots and LdSlots.

Fix it by looking for a label before parsing as an expression statement.
Labels are by spec exactly a tkID followed by a tkColon.  Chakra
unfortunately allows parenthesized label names, so we also check for
balanced open and close parentheses here to maintain that behavior.  We
will remove this legacy behavior in a future release.

If we don't find a label pattern in the syntax then we rewind and parse
as an expression statement.