Simplify label parsing and delete label pnode
knopLabel was not actually used even though there were no-op references
to it in the bytecode generator. knopLabel nodes were not ever placed
into the AST by the parser.  This appears to be a result of the
introduction of StmtNest in the parser during ES6 let/const development.

Label parsing was using knopLabel to track labels in StmtNests for
buildAST parses, and LabelId* lists for !buildAST.  This was redundant
and label parsing need only be implemented as one or the other.

This change removes the knopLabel node type and uses LabelId* lists
instead to track labels in both buildAST and !buildAST parsing.

This takes care of the following TODO in Parse.cpp:

    TODO[ianhall]: Refactor to eliminate separate code paths for
    buildAST and !buildAST