Fix handling of inlined ArgIn_Rest while mapping formals during inlining
While mapping actuals during inlining, do not handle ArgIn_Rest when the
instruction was inlined originally in the inlinee. Existing mapping code was
resulting in incorrect codegen when mapping ArgIn_Rest instruction which was
originally inlined into the inlinee. Because the mapping code queried
formalCount of the inlinee instead of the originally inlined function of
which ArgIn_Rest was part of, and created an incorrect size NewScArray.