Fix jitted initialization of StackScriptFunction.
(This is a fix that Rajat implemented and I'm pushing through in his
absence.) The -oopjit- code gen for StackScriptFuntion initialization assumes that
the deferred prototype type can be loaded directly from the FunctionProxy
that is available at JIT time. The problem is that the proxy may be
redeferred and reallocated, and the type pointer may now be referring to
something entirely different. To solve this, generate the same code for
OOPJIT and -oopjit-, reloading the FunctionProxy from the parent's nested array at
execution time.