Adding func addr range cache on script context for IsNativeAddr (WER case)
This is to enable WER scenario to perform IsNativeAddress check with OOPjit case.
IsNativeAddress in DAC wasn't handling OOPjit case correctly today.

Fix:
Since we will not be able to make RPC call during WER, we cache the func ranges on the script context and check this cache during IsNativeAddress.