Update totalOpHelperLengthByEnd on lifetime accurately for unstructured control flow.
We can encounter loops such as the following:

L1: // loop L1 header
...
jmp L1
L2: // loop L2 header
...
jmp L1
...
jmp L2

The outer loop (L1) lexically ends before the inner loop (L2) here. The extended lifetimes in
such outer loops are extended until the end of inner loop. Also, the
loopEnd of the outer loop is extended till the end of the inner loop.
However fields such as totalOpHelperLengthByEnd on the outer loop are not
updated.
In this change, we loop over the parents of curLoop which originally ended
lexically before the curLoop and update totalOpHelperLengthByEnd on its
extended lifetimes and other fields.