[MERGE #2455 @satheeshravi] Stack args fix for detecting scope obj restore
Merge pull request #2455 from satheeshravi:StackArgsBailoutFix_release1.4

The flag that is used for detecting if we restored scope object in the bail out path, was wrongly being marked while restoring outparams for .call bailout scenarios.

Fix:
Moved the setter to a point where we actually set the localClosure property with the scope object.