Fix lowering of BrOnNotEmpty
When we finish iterating the ForInCache in JIT'ed code and if the cache is completed, we can exit the loop.
We always generate BrOnEmpty in the byte code and it doesn't get inverted to BrOnNotEmpty usually.
But for BrOnNotEmpty, the target of the branch is the loop body instead of loop end.  We need to use the label after the instruction when the branch inversion happens.