Remove default arguments from member definitions
Default arguments are not permitted on out-of-line definitions of members
of class templates. They can only be used in the declaration in the class.

This is a pre-emptive fix for upcoming MSVC compiler changes that will
diagnose this code. The Clang build only allows these currently
because they are built with -fms-extensions -Wno-microsoft.