Fix for assertion fired in JSON.stringify.
We called the GetLength twice assuming that the length will not change. However with proxy you can give different length both time and that will causes an assert. No fre impact so just removed the assert. While testing I found that we can pass more that 2^32 and that will cause the overflow in the length. However the CompoundString cannot handle more than MaxCharLength (INT_MAX-1) and it will eventually OOM. Instead of waiting till that point I am raising the RangeError.