Fixing !ldsym in jd for OOPjit
We don't register the native addresses and interpreter thunk addresses for OOPjit scenarios - since the allocations linked list is not populated by the runtime process.

Fix:
Used the cache that we now populate on the script context to register the address.
For interpreter thunk emitter - we use the thunkBlocks linked list.