Fix parsing a lambda consisting of a single statement wrapped in parens
If we parse a lambda which only has a single statement which is wrapped
in parens but not curly braces, we calculate the extent of that lambda
parse node incorrectly.

var a = () => (123)

In above, the parse node for the lambda has it's limit character set to
the limit character of the return node - which is the numeric constant
123. The return node throws away the parens and, thus, the character
extents for the lambda as a whole are incorrectly pointing to the last
character of 123.

This is not a regression in behavior. In existing code, if you try to
toString a lambda like this it will get cutoff:

print('"' + a + '"'); // "() => (123"

However, after improving the utf8 parsing such that we correctly
calculate the extents of the bytes we need to parse, toString of the
lambda will hit an assert due to the mismatch in expected length and
actual characters parsed.

Fix here is to keep track of the last right paren we parsed when we're
parsing the lambda body expression and use that instead of the character
extent of the return parse node.