VSO 8294189: Fix two incorrect, high-hitting asserts in RegexHelper::StringReplace.
In the function

```
RegexHelper::StringReplace(JavascriptString* match, JavascriptString* input, JavascriptFunction* replacefn)
```

These two asserts would fail when `replacefn` is cross-site (where `scriptContext` is the context of the `replacefn`):

```
Assert(match->GetScriptContext() == scriptContext);
Assert(input->GetScriptContext() == scriptContext);
```

These asserts were overzealous. There is no actual problem here because the `replacefn`'s `ScriptContext` is not used for any outputs from this function and therefore is not observable.

The `scriptContext` in which `EntryReplace` is called should match `match->GetScriptContext()` and `input->GetScriptContext()`.

The fix for these asserts is to pass in the `scriptContext` from the calling function and check against that `scriptContext`. It is okay for the `scriptContext` of `replacefn` to be different, as the parameters and results will be marshalled correctly in other code.

Replaces #2458