Unshift should always do the precheck
Since the unshift expand to new length. It is possible that the current
array does not have any missing item but it still has accessor on the
proto chain. That setter will not be fired as we were taking the no-side
effect path. Fixed that by forcing to check the proto chain for the
unshift case.
Added test case for more understanding of the scenario.