Issue2681/2702: Multiple imports of same module
While resolving a module, a counter ("numUnInitializedChildModule")
is maintained to reflect the number of its child modules that have
not been instantiated. Once a child module is parsed and all its children
resolved, its parent modules will be notified through OnChildModuleReady(),
where the counter is decremented. The counter has to be zero for a module
to be instantiated.

This causes a problem when there are multiple imports referring to
the same child module, and the child module is parsed and resolved
before some of its other parent modules. Because the child module is only parsed
once, parent modules that are parsed and linked to the child module after
the notification will never see their counter going back to zero. As a result,
these parent modules and all modules importing them will never be instantiated.

Proposed fix is to add a flag ("readyAsChildModule") and flip it from false to true
upon notification of a module's parents. Future parent modules will see the
flag and deem that child module resolved. Also proposed a name change of the
counter from "numUnInitializedChildModule" to "numPendingChildModule" to
reflect the logic change.