Several changes to runtests.py
Remove the flag-based test ignoring of -simdjs as it has
caused confusion on whether or not tests were running.
Add test tagging for simd, asmjs, and wasm functionality.
Add the ability to read --sanitize flags and exclude known
failing tests.

Edit: Allow all of the asmjs tests, since they work now.

@obastemur fixed the issues with them in #2716, so I've gone and
enabled them here.

Edit: Add --static to test flags.

Not using it currently to mask out any tests; this is just to make
the test and build flags line up better.

Edit: Add a -j flag to runtests.py for num processes

This makes it line up better with the build flags for CI

Edit: Add a flag to warn instead of fail on test timeout

This is in an attempt to reduce noise a bit; a small number of tests
fail to run quickly sometimes enough on one channel of our test infra.

Edit: Specify type for -j test flag.

Edit: Disable tests that do not currently work with ASAN

Edit: Make the verbose flag do something, add a way to run tests outside
of the repo if needed.