Change WScriptJSRT LoadScript to actually free
Now calls to it pass a finalizeCallback, which is here
free, but could be more complex if the source text is
differently allocated.

Edit: Make calling conventions match up for free callback.

The calling conventions used are dependent on platform; change to
using a wrapper around free such that we always match the expected
convention.

Edit: Omit unneeded 'return' in free callback.