We do stack walking for this, so don't ASAN
One of the very few cases where ASAN reports a
false positive is in stack walking code, so go
and disable that here.

Edit: Create NO_SANITIZE_ADDRESS macro and use it

This is mostly a matter of removing duplicate attribute definitions
and making a macro that puts it in if needed.