ES6module cyclic imports fix and test enhancements
	1) ES6module: cyclic imports (Issue#2800)

(Reference: Issue2681/2702)

For ES6 module support, module import and resolution depends on a counter ("numPendingChildModule") to increment when a child module is identified and decrement when the child module "notifies" its parent module.

This breaks down when there imports forms a cycle, e.g. `mod0 > mod1 > mod2 > mod0` (">" represents left-side module imports right-side module). In this case, the counter stays at one indefinitely and none of these modules involved is instantiated.

Proposed fix is to break the cycle by checking the "wasParsed" flag of the child module, changed from false to true right after the parser has done processing the module code successfully. Once the "wasParsed" flag is set to "true", the child module will either resolve all its dependencies and be ready for instantiation, or find itself/dependencies in error state. Both cases will cause its parent module to be notified and decrement the counter, except when loading fails in the host.

The "readyAsChildModule" flag that indicates parent notification is renamed to "parentsNotified" and retained for assertion purpose.

	2) Unit test framework improvement for module

Cyclic imports were nominally covered in unit tests but did not catch Issue#2800. This is because module-related unit tests use WScript.LoadModule() that failed silently when modules fail to instantiate. We need an update on unit test framework to collect and tally results from tests that are launched asynchronously.

Proposed changes to the unit test framework augments testRunner APIs by 1) adding tracking of asynchronously launched code in runTest(); 2) adding testRunner.LoadModule() as a promise-friendly wrapper for Wscript.LoadModule(). Module tests therefore go through the extra step of checking asynchronously launched code, which have to complete with no error in order for the tests to pass.

	3) ModuleNamespace initialization triggers singleton instance assertion (Issue #2840)

Instances of ModuleNamespace share the same type object moduleNamespaceType in JavascriptLibrary. Need to mark the same as shared to avoid singleton instance logic in type handler classes.

	4) Tracing for module

This changeset also includes tracing and assertions for module.