Update return symbol assignment in the presence of early return in try finally
Currently we skip generating a Ld_A after LdElem for return symbol s0 to save a temporary register.
In the presence of try finally with an early return, we have to run the finally code after the function return.
This can lead to problems later on in globopt, this change fixes this.