Avoid adding empty blocks while inserting an edge from finally block to early exit
When we add an edge from end of finally block to an early exit, we need not create additional blocks
if the Leave/LeaveNull at the end of finally was in a separate block preceeded by a Label.
Creating empty blocks like this can lead to issues with DelateLeaveChains later on, moreover we want to keep the IR clean.