Adding the ability to dump perf.map files on non-windows systems
To enable tools such as `perf` to provide diagnostic information on
generated code, we add the ability to dump out a mapping from memory
addresses to function names. Sending a `SIGUSR2` signal to a ChakraCore
process on Linux or OSX will cause it to write to `/tmp/perf-<pid>.map`
the next time the process performs a stack probe.