Duplicate let or const definition in the body of a split scope funciton should throw redeclaration error
For non-split scope functions we add an additional reference to the body for each param symbol. When a declaration occurs in the body we will find that sym already existing and throw redeclaration. But in case of split scope functions we can't add the additional reference, so we have to check explicitly for duplicates.