When clearing inline cache remove it from invalidation list
During non-shutdown scenarios clearing the inline cache zeroes the
invalidationListSlotPtr but the node in various inlinecachelist on
threadcontext still point to inlinecache. In release build we only check
for values which are null and compact them but in debug build we have
validation which access the type stored thus reading invalid memory.
Invalidation should also help in more compaction.