Remove defunct tests.
These tests had been disabled for a very long time. The traces used
diverged significantly since then, and they're still not consistent
across different platforms or even necessarily machines on the same
platform. The tests don't seem to offer anything in excess of other
tests in the same directory, and they're too liable to break with a
change that does not impact functionality or perf.

Additionally, most editors and other tools 1-index lines. I changed
runtests.py to reflect this (it was printing 0-indexed line numbers
in test results).